Module heim_common::prelude

Prelude intended to be used across heim-* crates.

Consider not to use it in your code, because it is kinda internal and might change at any time.

Re-exports

pub use super::wrap;

pub use self::futures::future;

pub use self::futures::stream;

Modules

futures

This module tries to mimic futures crate structure except without re-exporting unused subcrates like executor or compat.

Structs

Error

Error type for data fetching operations.

Traits

FusedFuture

A future which tracks whether or not the underlying future should no longer be polled.

FusedStream

A stream which tracks whether or not the underlying stream should no longer be polled.

Future

A future represents an asynchronous computation.

FutureExt

An extension trait for Futures that provides a variety of convenient adapters.

Stream

A stream of values produced asynchronously.

StreamExt

An extension trait for Streams that provides a variety of convenient combinator functions.

TryFutureExt

Adapters specific to Result-returning futures

TryStreamExt

Adapters specific to Result-returning streams

Type Definitions

BoxFuture

An owned dynamically typed Future for use in cases where you can’t statically type your result or need to add some indirection.

BoxStream

An owned dynamically typed Stream for use in cases where you can’t statically type your result or need to add some indirection.

Result

A specialized Result type for data fetching functions.